Split thread utils from commanderbot.lib.forums
